put the navbar into separate component
